Switch to the LiveKit javascript SDK
I was having trouble getting the cloud function to
import the livekit sdk so made an endpoint with Cloud
Run in Rust that used the Rust LiveKit SDK.

Qais gave me a hand and we got the JS SDK to import.

I removed the axios dependency and the call to the endpoint.

The API Key and secret are stored in a .env file that is gitignored.